phase3/backend/requirements.txt
aiosqlite>=0.22.1
alembic>=1.17.2
apscheduler>=3.11.2
asyncpg>=0.31.0
bcrypt==4.2.1
email-validator>=2.3.0
fastapi>=0.128.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.12.0
python-dateutil>=2.9.0.post0
python-dotenv
python-jose[cryptography]
python-multipart
sqlmodel
uvicorn

# Phase III: AI Chatbot
openai
mcp